ocr_py_capture/requirements.txt
MouseInfo==0.1.3
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python3-xlib==0.15
pytweening==1.0.7
